karpathy-wiki/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karpathy-wiki"
version = "0.1.0"
description = "LLM Knowledge Base CLI implementing Karpathy's compilation pattern"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "ProtoGen"}
]
dependencies = [
    "typer[all]>=0.9.0",
    "boto3>=1.28.0",
    "watchdog>=3.0.0",
    "pypdf2>=3.0.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.scripts]
kw = "kw.cli:app"

[tool.setuptools]
packages = ["kw"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
